Drop dead helper, fix needless cast
- Remove unused `is_ok_or_dup` and its `StatusCode` import — leftover
  from before `UploadOutcome` carried the Created/Duplicate distinction
  on the typed result.
- Drop `secs as i64` cast where the lhs is already `time_t` (= i64 on
  Linux x86_64). clippy flagged.

`cargo clippy --no-deps` now clean.

+++ src/camera/gphoto.rs
@@ -44,7 +44,7 @@ pub fn digest_info(info: &FileInfo, filename: &str, tz: Tz) -> Result<ObjectInfo
 
 fn local_clock_secs_to_utc(secs: libc::time_t, tz: Tz) -> Result<DateTime<Utc>> {
     use chrono::TimeZone;
-    let pseudo_utc = DateTime::<Utc>::from_timestamp(secs as i64, 0)
+    let pseudo_utc = DateTime::<Utc>::from_timestamp(secs, 0)
         .ok_or_else(|| anyhow!("invalid timestamp"))?;
     let naive = pseudo_utc.naive_utc();
     let local = tz
